+++ doc/release/release_notes.rst
@@ -7,7 +7,7 @@
 Release Notes
 *************
 
-.. release:: upcoming
+.. release:: Upcoming
 
     .. change:: new